Moved xml_store_config, mongo_store_config, and draft_mongo_store_config to xmodule/modulestore/tests/django_util.py Removed some repeated instances of xml_store_config. Reverted some changes to mongo_store_config so each modulestore gets it's own uuid.
