+++ lms/djangoapps/courseware/tests/test_masquerade.py
@@ -15,7 +15,7 @@ from django.core.urlresolvers import reverse
 from django.contrib.auth.models import Group, User
 from courseware.access import _course_staff_group_name
 from courseware.tests.helpers import LoginEnrollmentTestCase
-from courseware.tests.modulestore_config import TEST_DATA_XML_MODULESTORE
+from modulestore_config import TEST_DATA_XML_MODULESTORE
 from xmodule.modulestore.django import modulestore
 import xmodule.modulestore.django
 import json
