+++ lms/djangoapps/courseware/tests/test_views.py
@@ -14,26 +14,11 @@ from xmodule.modulestore.django import modulestore
 import courseware.views as views
 from xmodule.modulestore import Location
 from pytz import UTC
+from modulestore_config import TEST_DATA_XML_MODULESTORE
 
 
 class Stub():
     pass
-
-
-# This part is required for modulestore() to work properly
-def xml_store_config(data_dir):
-    return {
-        'default': {
-            'ENGINE': 'xmodule.modulestore.xml.XMLModuleStore',
-            'OPTIONS': {
-                'data_dir': data_dir,
-                'default_class': 'xmodule.hidden_module.HiddenDescriptor',
-            }
-        }
-    }
-
-TEST_DATA_DIR = settings.COMMON_TEST_DATA_ROOT
-TEST_DATA_XML_MODULESTORE = xml_store_config(TEST_DATA_DIR)
 
 
 @override_settings(MODULESTORE=TEST_DATA_XML_MODULESTORE)
